requirements.txt
requests
pyautogui
botcity-core